requirements.txt
absl-py==1.3.0
action-tutorials-py==0.9.4
ament-copyright==0.9.7
ament-cppcheck==0.9.7
ament-cpplint==0.9.7
ament-flake8==0.9.7
ament-index-python==1.1.0
ament-lint==0.9.7
ament-lint-cmake==0.9.7
ament-package==0.9.5
ament-pep257==0.9.7
ament-uncrustify==0.9.7
ament-xmllint==0.9.7
asttokens==2.2.1
astunparse==1.6.3
backcall==0.2.0
cachetools==5.2.0
certifi==2022.12.7
charset-normalizer==2.1.1
comm==0.1.2
contourpy==1.0.6
cycler==0.11.0
dearpygui==1.8.0
debugpy==1.6.4
decorator==5.1.1
demo-nodes-py==0.9.4
domain-coordinator==0.9.2
entrypoints==0.4
examples-rclpy-executors==0.9.4
examples-rclpy-minimal-action-client==0.9.4
examples-rclpy-minimal-action-server==0.9.4
examples-rclpy-minimal-client==0.9.4
examples-rclpy-minimal-publisher==0.9.4
examples-rclpy-minimal-service==0.9.4
examples-rclpy-minimal-subscriber==0.9.4
executing==1.2.0
flatbuffers==22.12.6
fonttools==4.38.0
gast==0.4.0
google-auth==2.15.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.51.1
h5py==3.7.0
idna==3.4
importlib-metadata==5.1.0
ipykernel==6.19.2
ipython==8.7.0
jedi==0.18.2
joblib==1.2.0
jupyter_client==7.4.8
jupyter_core==5.1.0
keras==2.11.0
kiwisolver==1.4.4
launch==0.10.10
launch-ros==0.11.7
launch-testing==0.10.10
launch-testing-ros==0.11.7
launch-xml==0.10.10
launch-yaml==0.10.10
libclang==14.0.6
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.6.2
matplotlib-inline==0.1.6
nest-asyncio==1.5.6
numpy==1.23.5
oauthlib==3.2.2
opt-einsum==3.3.0
osrf-pycommon==0.1.11
packaging==22.0
pandas==1.5.2
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.3.0
pkg_resources==0.0.0
platformdirs==2.6.0
prompt-toolkit==3.0.36
protobuf==3.19.6
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.13.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.6
pyzmq==24.0.1
quality-of-service-demo-py==0.9.4
requests==2.28.1
requests-oauthlib==1.3.1
ros2action==0.9.12
ros2bag==0.3.9
ros2cli==0.9.12
ros2component==0.9.12
ros2doctor==0.9.12
ros2interface==0.9.12
ros2launch==0.11.7
ros2lifecycle==0.9.12
ros2multicast==0.9.12
ros2node==0.9.12
ros2param==0.9.12
ros2pkg==0.9.12
ros2run==0.9.12
ros2service==0.9.12
ros2topic==0.9.12
rosidl-runtime-py==0.9.1
rpyutils==0.2.0
rqt-action==0.4.9
rqt-console==1.1.2
rqt-graph==1.1.3
rqt-gui==1.1.2
rqt-gui-py==1.1.2
rqt-msg==1.0.5
rqt-plot==1.1.1
rqt-publisher==1.3.0
rqt-py-console==1.0.2
rqt-reconfigure==1.0.8
rqt-service-caller==1.0.5
rqt-shell==1.0.2
rqt-srv==1.0.3
rqt-top==1.0.2
rqt-topic==1.3.0
rsa==4.9
scikit-learn==1.2.0
scipy==1.9.3
six==1.16.0
sros2==0.9.5
stack-data==0.6.2
tdqm==0.0.1
teleop-twist-keyboard==2.3.2
tensorboard==2.11.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.11.0
tensorflow-estimator==2.11.0
tensorflow-io-gcs-filesystem==0.28.0
termcolor==2.1.1
threadpoolctl==3.1.0
topic-monitor==0.9.4
tornado==6.2
tqdm==4.64.1
traitlets==5.7.1
turtlebot3-example==2.1.1
turtlebot3-teleop==2.1.1
typing_extensions==4.4.0
urllib3==1.26.13
wcwidth==0.2.5
Werkzeug==2.2.2
wrapt==1.14.1
zipp==3.11.0
